Restructure pageSelector under Pillars component
