+++ src/components/Layout.js
@@ -3,7 +3,7 @@ import { Container, Row, Col, Hidden, Visible } from 'react-grid-system';
 import Logo from './Logo';
 import Header from './header/Header';
 import AboutPage from './about/AboutPage';
-import PageSelector from './body/pageSelector/PageSelector'
+import Pillars from './body/pillars/Pillars';
 import Footer from './footer/Footer';
 
 class Layout extends Component {
@@ -17,7 +17,7 @@ class Layout extends Component {
           <div className="container div-center">
             <AboutPage />
             <hr />
-            <PageSelector />
+            <Pillars />
             {this.props.children}
           </div>
